TEST: [cmn] Refactor recent net modules' testing code
