+++ crates/common/src/proxy/proxy_channel_and_tcp.rs
@@ -368,7 +368,7 @@ unsafe impl Send for ChannelAndTcpStreamProxy {}
 #[cfg(test)]
 pub mod tests {
     use super::*;
-    use crate::net::stream_utils::tests::ConnectedTcpStream;
+    use crate::net::stream_utils::ConnectedTcpStream;
     use anyhow::Result;
     use std::io::ErrorKind::WouldBlock;
     use std::io::{Read, Write};
